Fixed last of the warnings from g++.

+++ lib/spipe/lib/sslib/src/analysis/ConvexHull.cpp
@@ -343,7 +343,7 @@ ConvexHull::generateHullPoint(const common::AtomsFormula & composition, const Hu
   ::std::vector< HullTraits::FT> tempVec(myHullDims);
   VectorD v(myHullDims);
   int numEndpoints = 0;
-  for(int i = 0; i < myEndpoints.size(); ++i)
+  for(size_t i = 0; i < myEndpoints.size(); ++i)
   {
     numAtoms = composition.wholeNumberOf(myEndpoints[i].first);
     if(numAtoms != 0)
@@ -483,7 +483,7 @@ ConvexHull::canGenerate() const
 void
 ConvexHull::initEndpoints(const EndpointLabels & labels)
 {
-  for(int i = 0; i < labels.size(); ++i)
+  for(size_t i = 0; i < labels.size(); ++i)
   {
     myEndpoints.push_back(::std::make_pair(labels[i], PointD(myHullDims)));
     // Make sure we're using the reduced, i.e. most simple, formula
